Cleanup, moving stuff to db package, setting up for cache syncing
