+++ gateway/db/accounts.go
@@ -1,25 +1,13 @@
-package proxy
+package db 
 
 import (
     "context"
     "fmt"
-    "github.com/redis/go-redis/v9"
     "strconv"
-    "sync"
 )
 
-var client *redis.Client = nil
-var redisMutex sync.Once
-
-func getClient() *redis.Client {
-    redisMutex.Do(func() {
-        client = redis.NewClient(&redis.Options{
-            Addr: "localhost:6379",
-            Password: "",
-            DB: 0,
-        })
-    })
-    return client
+type Account struct {
+    id string
 }
 
 func IncRelayCount(ctx context.Context, account string) {
@@ -34,7 +22,7 @@ func IncRelayCount(ctx context.Context, account string) {
 }
 
 // TODO remove, just for testing
-func IncCounter(ctx context.Context, name string) int64 {
+func IncrCounter(ctx context.Context, name string) int64 {
     result, err := getClient().Incr(ctx, name).Result()
     if err != nil {
         // TODO handle errors properly
@@ -43,7 +31,7 @@ func IncCounter(ctx context.Context, name string) int64 {
     return result
 }
 
-func DecCounter(ctx context.Context, name string) int64 {
+func DecrCounter(ctx context.Context, name string) int64 {
     result, err := getClient().Decr(ctx, name).Result()
     if err != nil {
         // TODO handle errors properly
@@ -72,4 +60,16 @@ func GetIntVal(ctx context.Context, name string) int {
         return 0
     }
     return intval
+}
+
+// TODO check redis, if missing check postgres (and cache), if neither return false
+func LookupAccount(ctx context.Context, apiKey string) (Account, bool) {
+    // TODO do this in tx that checks and gets account information?
+    result, err := getClient().SIsMember(ctx, ACCOUNT_SET, apiKey).Result()
+    if err != nil || !result {
+        // TODO handle errors better
+        return Account{}, false
+    }
+    // TODO need to grab all keys to build account
+    return Account{""}, true
 }
